Promoted MakeResourceRequest
to dispatch to a Resource, based upon the method
