+++ playground/routes.go
@@ -2,10 +2,8 @@ package playground
 
 import (
 	"io"
-	"sort"
 
 	"github.com/kkrull/gohttp/http"
-	"github.com/kkrull/gohttp/msg/clienterror"
 )
 
 func NewReadOnlyRoute() *ReadOnlyRoute {
@@ -21,7 +19,7 @@ func (route *ReadOnlyRoute) Route(requested *http.RequestLine) http.Request {
 		return nil
 	}
 
-	return lookupRequest(requested, route.Resource)
+	return http.MakeResourceRequest(requested, route.Resource)
 }
 
 type ReadOnlyResource interface {
@@ -44,7 +42,7 @@ func (route *ReadWriteRoute) Route(requested *http.RequestLine) http.Request {
 		return nil
 	}
 
-	return lookupRequest(requested, route.Resource)
+	return http.MakeResourceRequest(requested, route.Resource)
 }
 
 type ReadWriteResource interface {
@@ -52,57 +50,4 @@ type ReadWriteResource interface {
 	Head(client io.Writer)
 	Post(client io.Writer)
 	Put(client io.Writer)
-}
-
-func lookupRequest(requested *http.RequestLine, resource interface{}) http.Request {
-	if requested.Method == "OPTIONS" {
-		return &knownOptionsRequest{
-			SupportedMethods: supportedMethods(requested.Target, resource),
-		}
-	}
-
-	method := knownMethods[requested.Method]
-	if method == nil {
-		return unknownHttpMethod(requested, resource)
-	}
-
-	request := method.MakeRequest(requested, resource)
-	if request == nil {
-		return unsupportedMethod(requested, resource)
-	}
-
-	return request
-}
-
-func unknownHttpMethod(requested *http.RequestLine, resource interface{}) http.Request {
-	return clienterror.MethodNotAllowed(supportedMethods(requested.Target, resource)...)
-}
-
-func unsupportedMethod(requested *http.RequestLine, resource interface{}) http.Request {
-	return clienterror.MethodNotAllowed(supportedMethods(requested.Target, resource)...)
-}
-
-func supportedMethods(target string, resource interface{}) []string {
-	supported := []string{"OPTIONS"}
-	for name, method := range knownMethods {
-		imaginaryRequest := &http.RequestLine{Method: name, Target: target}
-		request := method.MakeRequest(imaginaryRequest, resource)
-		if request != nil {
-			supported = append(supported, name)
-		}
-	}
-
-	sort.Strings(supported)
-	return supported
-}
-
-var knownMethods = map[string]Method{
-	"GET":  &getMethod{},
-	"HEAD": &headMethod{},
-	"POST": &postMethod{},
-	"PUT":  &putMethod{},
-}
-
-type Method interface {
-	MakeRequest(requested *http.RequestLine, resource interface{}) http.Request
 }
